feat: add mock implementation for MongoDB operations
This commit adds a new file `mongo/mock.go` which contains a mock implementation for various MongoDB operations. This mock implementation is useful for testing purposes, allowing us to simulate the behavior of a real MongoDB client without actually interacting with a live database.

The `MockMongoOperations` struct in the `mongo` package now includes methods such as `GetMongoClient`, `GetMongoDatabase`, `GetMongoCollection`, `InsertOne`, `InsertMany`, `FindOne`, `Find`, `UpdateOne`, `UpdateMany`, `DeleteOne`, and `DeleteMany`. These methods return mocked results and errors based on the input parameters.

The purpose of this mock implementation is to facilitate unit testing of code that depends on MongoDB operations. By using the mock implementation, we can isolate the code under test and verify its behavior without the need for a live MongoDB server.

This commit also includes some minor formatting and import changes in the `mongo/mock.go` file.
